src/cloud_run/mimitalk_calc_dev/requirements.txt
# Cloud Functionsの実行に必須のフレームワーク
functions-framework==3.*

# Vertex AI Agentと通信するために必要 (vertexaiライブラリ)
google-cloud-aiplatform>=1.49.0

# Firestoreデータベースから会話ログを読み込むために必要
google-cloud-firestore